feat: refactor code to process requests in multiple threads
