+++ package.json
@@ -12,16 +12,16 @@
     "@bundlr-network/client": "0.11.6",
     "arweave": "^1.13.0",
     "bundle-require": "^4.0.1",
-    "esbuild": "^0.17.8",
     "graphql": "^16.6.0",
     "pino": "^8.12.1",
-    "poolifier": "^2.6.21",
     "request": "^2.88.2",
     "warp-contracts": "^1.4.14",
-    "warp-contracts-plugin-deploy": "^1.0.9"
+    "warp-contracts-plugin-deploy": "^1.0.9",
+    "workerpool": "^6.4.0"
   },
   "devDependencies": {
     "@types/node": "^18.15.11",
+    "@types/workerpool": "^6.4.0",
     "@typescript-eslint/eslint-plugin": "^5.59.1",
     "@typescript-eslint/parser": "^5.59.1",
     "eslint": "^8.39.0",
